requirements.txt
fastapi>=0.108.0
uvicorn[standard]>=0.24.0,<0.27.0
sqlalchemy>=2.0.25,<3.0.0
pydantic>=2.5.3,<3.0.0
jinja2>=3.1.2,<4.0.0
python-multipart==0.0.6
python-dateutil>=2.8.2
reportlab>=4.0.7
openpyxl>=3.1.2
pandas>=2.1.4
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
pillow>=10.1.0

# Autenticação e Segurança
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
email-validator>=2.2.0
